network/requirements.txt
django==5.2.4
djangorestframework==3.16.0
pillow==11.3.0
psycopg2-binary==2.9.10
geopy==2.4.1
python-dotenv==1.1.1
geopy==2.4.1